Fix navigation configuration and file extension
